Rename BufferView -> Editor
* BufferElement -> EditorElement
